requirements.txt
sqlalchemy==1.3.22
requests==2.25.1
PyJWT==2.0.0
discord.py==1.6.0
python-dotenv==0.15.0
raygun4py==4.3.0